Added youtube import option

+++ app/Http/Controllers/ImportController.php
@@ -6,10 +6,8 @@ use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\DB;
+use App\Services\ImportService;
 use Illuminate\Http\Request;
-use App\Models\TextBlock;
-use App\Models\Lesson;
-use App\Models\Book;
 
 
 class ImportController extends Controller
@@ -18,11 +16,17 @@ class ImportController extends Controller
     public function import(Request $request) {
         $userId = Auth::user()->id;
         $importType = $request->post('importType');
-        $importFile = $request->file('importFile');
         $textProcessingMethod = $request->post('textProcessingMethod');
         $bookId = $request->post('bookId');
         $bookName = $request->post('bookName');
         $chapterName = $request->post('chapterName');
+        
+        if ($importType == 'e-book') {
+            $importFile = $request->file('importFile');
+        } else {
+            $importText = $request->post('importText');
+        }
+        
 
         // move file to temp folder
         if ($importType === 'e-book') {
@@ -32,12 +36,21 @@ class ImportController extends Controller
             $importFile->move(storage_path('app/temp'), $fileName);
         }
 
-        // import text
+        // import
         try {
-            $this->importBook($textProcessingMethod, storage_path('app/temp') . '/' . $fileName, $bookId, $bookName, $chapterName);
+            if ($importType === 'e-book') {
+                // e-book
+                (new ImportService())->importBook($textProcessingMethod, storage_path('app/temp') . '/' . $fileName, $bookId, $bookName, $chapterName);
+            } else {
+                // text
+                (new ImportService())->importText($textProcessingMethod, $importText, $bookId, $bookName, $chapterName);
+            }
         } catch (\Exception $exception) {
-            File::delete(storage_path('app/temp') . '/' . $fileName);
-            return 'error';
+            if ($importType === 'e-book') {
+                File::delete(storage_path('app/temp') . '/' . $fileName);
+            }
+
+            return $exception->getMessage();
         }
 
         // delete temp file
@@ -48,80 +61,10 @@ class ImportController extends Controller
         return 'success';
     }
 
-    private function importBook($textProcessingMethod, $file, $bookId, $bookName, $chapterName) {
-        DB::disableQueryLog();
-        $userId = Auth::user()->id;
-        $selectedLanguage = Auth::user()->selected_language;
+    public function getYoutubeSubtitles(Request $request) {
+        $url = $request->post('url');
+        $subtitleList = (new ImportService())->getYoutubeSubtitles($url);
 
-        // tokenize book
-        $text = Http::post('linguacafe-python-service:8678/tokenizer/import', [
-            'language' => $selectedLanguage,
-            'textProcessingMethod' => $textProcessingMethod,
-            'importFile' => $file,
-            'chunkSize' => 4000
-        ]);
-        
-        // get text and token chunks
-        $text = json_decode($text);
-        $chunks = $text->processedChunks;
-        $textChunks = $text->textChunks;
-
-        // retrieve or create book
-        if ($bookId == -1) {
-            $book = new Book();
-            $book->user_id = $userId;
-            $book->cover_image = 'default.jpg';
-            $book->language = $selectedLanguage;
-            $book->name = $bookName;
-            $book->save();
-        } else {
-            $book = Book
-                ::where('user_id', $userId)
-                ->where('id', $bookId)
-                ->first();
-            
-            if (!$book) {
-                return 'error';
-            }
-        }
-
-        // import each chunk as a chapter
-        foreach ($chunks as $chunkIndex => $chunk) {
-            $lesson = new Lesson();
-            $lesson->user_id = $userId;
-            $lesson->name = $chapterName . ' ' . ($chunkIndex + 1);
-            $lesson->read_count = 0;
-            $lesson->word_count = 0;
-            $lesson->book_id = $book->id;
-            $lesson->language = $selectedLanguage;
-            $lesson->raw_text = $textChunks[$chunkIndex];
-            $lesson->unique_words = '';
-            $lesson->setProcessedText([]);
-
-            $textBlock = new TextBlock();
-            $textBlock->tokenizedWords = $chunk;
-            $textBlock->processTokenizedWords();
-            $textBlock->collectUniqueWords();
-            $textBlock->updateAllPhraseIds();
-            $textBlock->createNewEncounteredWords();
-            
-            $uniqueWordIds = DB
-                ::table('encountered_words')
-                ->select('id')
-                ->where('user_id', Auth::user()->id)
-                ->where('language', $selectedLanguage)
-                ->whereIn('word', $textBlock->uniqueWords)
-                ->pluck('id')
-                ->toArray();
-
-            // update lesson word data
-            $lesson->setProcessedText($textBlock->processedWords);
-            $lesson->word_count = $textBlock->getWordCount();
-            $lesson->unique_words = json_encode($textBlock->uniqueWords);
-            $lesson->unique_word_ids = json_encode($uniqueWordIds);
-            $lesson->save();
-        }
-
-        return 'success';
+        return json_encode($subtitleList);
     }
 }
